1、普华释放1.16.1软件集成，bcm04 tms05 vcu06

+++ Gecko_E3420_CCU_R19_MCAL3_1_IAR_V8_50_6_Project_PPV/driver/RF/NJJ29C0B_Api.h
@@ -7,6 +7,6 @@
 
 extern void NJJ29C0B_Init(void);
 extern Dio_LevelType Get_NJJ29C0B_INT_PinLevel(void);
-extern Std_ReturnType NJJ29C0B_Frame_Transmit(uint8* payload_tx, uint8* rxdata, uint8 payload_len);
+extern Std_ReturnType NJJ29C0B_Frame_Transmit(uint8* payload_tx, uint8* rxdata, uint16 payload_len);
 
 #endif /*__NJJ29C0B_API_H_*/
